requirements.txt
fastapi
uvicorn
httpx
groq
python-dotenv
jinja2
python-multipart
